h100/requirements.txt
openai>=1.0.0
datasets>=2.20.0
Pillow>=10.0.0
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0
decord>=0.6.0
av>=10.0.0
huggingface_hub>=0.20.0
# BFCL-V4 web search (free, no API key needed)
duckduckgo-search>=6.0.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
requests>=2.31.0
